+++ src/tracing/mod.rs
@@ -394,7 +394,7 @@ impl TracingInspector {
         let trace_idx = self.pop_trace_idx();
         let trace = &mut self.traces.arena[trace_idx].trace;
 
-        trace.gas_used = gas.spent();
+        trace.gas_used = gas.total_gas_spent();
         trace.gas_refund_counter = gas.refunded().max(0) as u64;
 
         trace.status = Some(result);
@@ -467,7 +467,7 @@ impl TracingInspector {
 
         let gas_used = gas_used(
             interp.runtime_flag.spec_id(),
-            interp.gas.spent(),
+            interp.gas.total_gas_spent(),
             interp.gas.refunded() as u64,
         );
 
@@ -540,9 +540,15 @@ impl TracingInspector {
         if self.config.record_stack_snapshots.is_all()
             || self.config.record_stack_snapshots.is_pushes()
         {
-            // this can potentially underflow if the stack is malformed
-            let start = interp.stack.len().saturating_sub(step.op.outputs() as usize);
-            step.push_stack = Some(interp.stack.data()[start..].into());
+            let outputs = if step.op.is_valid() { step.op.outputs() as usize } else { 0 };
+            step.push_stack = Some(
+                interp
+                    .stack
+                    .data()
+                    .get(interp.stack.len().saturating_sub(outputs)..)
+                    .unwrap_or_default()
+                    .into(),
+            );
         }
 
         let journal = context.journal_ref().journal();